requirements.txt
requests==2.18.4
requests-mock==1.4.0
GitPython==2.1.8
prospector==0.12.7
pytest==3.4.0
